requirements.txt

Django==3.2.5
pylint==2.9.6
psycopg2==2.9.1
psycopg2-binary==2.9.1
djangorestframework==3.12.4
gunicorn==20.1.0
django-filter==2.4.0
dj-database-url==0.5.0
whitenoise==3.2.2
django-heroku==0.3.1
requests==2.26.0

